refactor: seperate ScaleList from index.tsx
